+++ setup.py
@@ -13,7 +13,7 @@ setup(name='pyTelegramBotAPI',
       author='eternnoir',
       author_email='eternnoir@gmail.com',
       url='https://github.com/eternnoir/pyTelegramBotAPI',
-      packages=['pytelegrambotapi'],
+      packages=['telebot'],
       license='GPL2',
       keywords='telegram bot api tools',
       install_requires=['requests', 'six'],
